wire up auth service
